Split device tests into a separate file
