+++ tests/components/mazda/test_sensor.py
@@ -1,6 +1,5 @@
 """The sensor tests for the Mazda Connected Services integration."""
 
-from homeassistant.components.mazda.const import DOMAIN
 from homeassistant.const import (
     ATTR_FRIENDLY_NAME,
     ATTR_ICON,
@@ -10,38 +9,10 @@ from homeassistant.const import (
     PERCENTAGE,
     PRESSURE_PSI,
 )
-from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers import entity_registry as er
 from homeassistant.util.unit_system import IMPERIAL_SYSTEM
 
 from tests.components.mazda import init_integration
-
-
-async def test_device_nickname(hass):
-    """Test creation of the device when vehicle has a nickname."""
-    await init_integration(hass, use_nickname=True)
-
-    device_registry = dr.async_get(hass)
-    reg_device = device_registry.async_get_device(
-        identifiers={(DOMAIN, "JM000000000000000")},
-    )
-
-    assert reg_device.model == "2021 MAZDA3 2.5 S SE AWD"
-    assert reg_device.manufacturer == "Mazda"
-    assert reg_device.name == "My Mazda3"
-
-
-async def test_device_no_nickname(hass):
-    """Test creation of the device when vehicle has no nickname."""
-    await init_integration(hass, use_nickname=False)
-
-    device_registry = dr.async_get(hass)
-    reg_device = device_registry.async_get_device(
-        identifiers={(DOMAIN, "JM000000000000000")},
-    )
-
-    assert reg_device.model == "2021 MAZDA3 2.5 S SE AWD"
-    assert reg_device.manufacturer == "Mazda"
-    assert reg_device.name == "2021 MAZDA3 2.5 S SE AWD"
 
 
 async def test_sensors(hass):
